--- Assets/Scripts/PlayerControler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b7a45a4d85b576e469ea49c884d41eb4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
