+++ Assets/Scripts/PlayerController.cs
@@ -1,0 +1,23 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerController : MonoBehaviour
+{
+    public float movementSpeed = 5f;
+    public Rigidbody2D rb;
+    Vector2 movement;
+    public Joystick joystick;
+
+    // Update is called once per frame
+    void Update()
+    {
+        movement.x = joystick.Horizontal;
+        movement.y = joystick.Vertical;
+    }
+
+    private void FixedUpdate()
+    {
+        rb.MovePosition(rb.position + movement * movementSpeed * Time.fixedDeltaTime);
+    }
+}
